Defactoring PullRequest
