+++ composeApp/src/androidMain/kotlin/band/effective/office/elevator/ui/UIPreview.kt
@@ -2,10 +2,9 @@ package band.effective.office.elevator.ui
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
-import band.effective.office.elevator.ui.employees.EmployeesScreen
 
 @Preview
 @Composable
 fun EmployeesScreenPreview(){
-    EmployeesScreen()
+    //EmployeeScreen()
 }
